Rename project to shelve.
Rename the project to shelve, since there is already a project named
stash listed on PyPI. The code is now licensed under a BSD license.
The directory structure has been revised to be in line with setuptools
standards. The project is now divided in a console script 'shelve.py',
and a package 'shelf'. The console script depends on functionality
offered in the package.