Refactor overall repository abstraction design.
Instead of a separate Shelf implementation for each repository, all
repository specific functionality is now wrapped in a class deriving
from Repository, which defines an interface for all repository
functionality this is needed by Shelf. This repository instance is then
used within the Shelf class to interact with the repository.

Furthermore, the exception class was moved to a separate module to
prevent import problems.